load module refactoring
