+++ morphs/__init__.py
@@ -1,5 +1,7 @@
 from __future__ import absolute_import
+import morphs.utils
 from morphs.utils import spikes, logistic
 from morphs.config import parallel, subj, paths
 import morphs.data
+import morphs.load
 import morphs.plot
