+++ morphs/data/__init__.py
@@ -1,3 +1,2 @@
 from __future__ import absolute_import
-import morphs.data.load
 import morphs.data.parse
