+++ morphs/data/localize.py
@@ -22,7 +22,7 @@ def _read_python(path):
 
 def calc_loc(block_path, squared=True):
     columns = ['block_path', 'AP', 'ML', 'Z', 'cluster', 'cluster_pos', 'cluster_accuracy']
-    spikes = morphs.data.load.ephys_data(block_path, collapse_endpoints=True)
+    spikes = morphs.load.ephys_data(block_path, collapse_endpoints=True)
     if len(spikes) == 0:
         return pd.DataFrame(columns=columns)
     waveforms, cluster_map = ephys.clust.compute_cluster_waveforms_fast(block_path, spikes)
@@ -45,7 +45,7 @@ def calc_loc(block_path, squared=True):
     i_cluster_map = {v: k for k, v in cluster_map.items()}
     d['cluster'] = [i_cluster_map[i] for i in range(len(cluster_map))]
     d['cluster_pos'] = y_hats
-    _, cluster_accuracies = morphs.data.load.cluster_accuracies()
+    _, cluster_accuracies = morphs.load.cluster_accuracies()
     d['cluster_accuracy'] = cluster_accuracies[block_path].loc[d['cluster']]['accuracy'].values
     return pd.DataFrame(data=d, columns=columns)
 
@@ -61,10 +61,10 @@ def generate_all_loc(parallel=False, n_jobs=morphs.parallel.N_JOBS):
     all_locs_df.to_pickle(morphs.paths.LOCATIONS_PKL)
 
 
-@morphs.data.load._create(morphs.paths.LOCATIONS_PKL, generate_all_loc, download_func=download_all_loc)
+@morphs.utils.load._create(morphs.paths.LOCATIONS_PKL, generate_all_loc, download_func=download_all_loc)
 def load_all_loc(prefer_download=True):
-    return morphs.data.load._pickle(morphs.paths.LOCATIONS_PKL.as_posix())
+    return morphs.utils.load._pickle(morphs.paths.LOCATIONS_PKL.as_posix())
 
 
 def download_all_loc():
-    morphs.data.load._download(morphs.paths.LOCATIONS_PKL, '1wLoMiKJjKPQbNF_qplqrMzHLyFCyFXn3')
+    morphs.utils.load._download(morphs.paths.LOCATIONS_PKL, '1wLoMiKJjKPQbNF_qplqrMzHLyFCyFXn3')
