+++ morphs/utils/__init__.py
@@ -1,0 +1,1 @@
+import morphs.utils.load
