+++ testing/test_accuracies.py
@@ -20,8 +20,8 @@ def test_cluster_accuracy():
     assert morphs.paths.ACCURACIES_PKL.exists()
     accuracies, cluster_accuracies = morphs.data.accuracies.load_cluster_accuracies()
     cluster = cluster_accuracies[block_path].index[-1]
-    spikes = morphs.data.load.ephys_data(block_path, good_clusters=[cluster],
-                                         collapse_endpoints=True)
+    spikes = morphs.load.ephys_data(block_path, good_clusters=[cluster],
+                                    collapse_endpoints=True)
     assert len(spikes['recording'].unique()) >= 1
     template_spikes = spikes[spikes['stim_id'].isin(list('abcdefgh'))]
     cluster_groups = template_spikes.groupby('cluster')
